requirements.txt
fastapi
uvicorn[standard]
solana
python-dotenv
SQLAlchemy
